Makes the library pip-installable and turns a master merge into a release: itbuilds, tests, and publishes wheels across Linux, macOS, and Windows.

**Packaging and distribution**
- Publish the Python bindings as `cartan-bindings`: dev wheels and sdist to TestPyPI from develop, production wheels and sdist to PyPI on master via OIDC trusted publishing.
- Derive the published version from `pyproject.toml`; suffix a unique dev version on develop and ship the verbatim version on master.
- Gate every built wheel on the full Python test suite, per interpreter and per platform, instead of a smoke test.
- Stop the tutorial tests from shadowing an installed wheel with the source tree on a clean checkout.

**Binding correctness**
- Format and parse numeric values locale-independently to stop a numpy segfault in the bindings.

**Wheel CI and platform matrix**
- Retier the wheel workflow: limited (Linux) on develop, full (Linux, macOS, Windows) on master and release PRs.
- Drop the manylinux2014 target and require glibc 2.28+.
- Cap the Python 3.10 NumPy compatibility sweep at the 2.2 series.

**Cross-platform build stability**
- Restore /EHsc and detect MSVC exceptions via _CPPUNWIND on Windows.
- Probe -Wno-c2y-extensions support instead of matching the compiler name.
- Supply Eigen to the ESP-IDF builds and correct the component manifest.
- Build CI native-only until argmin is embeddable via FetchContent.

**Tests**
- Make the projected_lm bitwise golden comparison opt-in and reference-build only, and keep it from running zero assertions.
- Cast float norms to double in the IK tolerance sweep and rename the angle-unwrap tolerance to avoid a POSIX name clash.

**Docs**
- Add the banner logo and public-preview status, and tidy the README.

+++ tests/unit/ik_float_tolerance_sweep_test.cpp
@@ -289,8 +289,8 @@ TEST_CASE("float IK converges at the swept preset but not at a forced 1e-6 gate"
 
         auto fk_sol = cartan::forward_kinematics(chain, result->solution.position);
         auto err = (fk_sol.end_effector.inverse() * target).log();
-        const double ori_err = err.head<3>().norm();
-        const double pos_err = err.tail<3>().norm();
+        const double ori_err = static_cast<double>(err.head<3>().norm());
+        const double pos_err = static_cast<double>(err.tail<3>().norm());
 
         std::printf(
             "[ik-float-conv] re-verified pose error: pos=%.3e m, ori=%.3e rad "
